embedding_evolution/requirements.txt
psycopg2-binary>=2.9.9
pgvector>=0.2.4
sqlalchemy>=2.0.23
pillow>=10.1.0
exifread>=3.0.0
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
sentence-transformers>=2.2.2
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
alembic>=1.12.0
